TracesConverter is now an interface.
